requirements.txt
absl-py==2.1.0
accelerate==1.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
anaconda-anon-usage @ file:///C:/b/abs_e8r_zga7xy/croot/anaconda-anon-usage_1732732454901/work
annotated-types @ file:///C:/b/abs_0dmaoyhhj3/croot/annotated-types_1709542968311/work
anyio==4.8.0
appdirs==1.4.4
archspec @ file:///croot/archspec_1709217642129/work
asgiref==3.11.1
async-timeout==4.0.3
attrs==25.1.0
autopep8==1.5.7
autotiktokenizer==0.2.2
beautifulsoup4==4.12.3
blinker==1.9.0
blis==0.7.11
boltons @ file:///C:/b/abs_45_52ughkz/croot/boltons_1737061711836/work
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
cachetools==5.5.1
catalogue==2.0.8
certifi==2025.1.31
cffi==1.14.5
chardet==5.2.0
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
chonkie==0.4.1
click==8.1.7
cloudpathlib==0.20.0
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
coloredlogs==15.0.1
colpali_engine==0.3.8
comtypes==1.4.10
conda @ file:///C:/b/abs_12ybmrg1d4/croot/conda_1738168389704/work
conda-anaconda-telemetry @ file:///C:/b/abs_4c9llcc5ob/croot/conda-anaconda-telemetry_1736524617431/work
conda-anaconda-tos @ file:///C:/b/abs_ceeuq0lee_/croot/conda-anaconda-tos_1739299022910/work
conda-content-trust @ file:///C:/b/abs_bdfatn_wzf/croot/conda-content-trust_1714483201909/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1737733694612/work/src
conda-package-handling @ file:///C:/b/abs_7fz3aferfv/croot/conda-package-handling_1731369038903/work
conda_package_streaming @ file:///C:/b/abs_bdz9vbvbh2/croot/conda-package-streaming_1731366449946/work
confection==0.1.5
contourpy==1.3.1
cryptography @ file:///C:/b/abs_e2lzchf4i6/croot/cryptography_1732130411942/work
cvzone==1.6.1
cycler==0.10.0
cymem==2.0.11
dataclasses-json==0.6.7
datasets==3.3.1
decorator==4.4.2
deepsearch-glm==1.0.0
deprecation==2.1.0
dill==0.3.8
diskcache==5.6.3
distro @ file:///C:/b/abs_71xr36ua5r/croot/distro_1714488282676/work
Django==4.2.29
Django-Verify-Email==1.0.7
docling==2.23.0
docling-core==2.19.1
docling-ibm-models==3.3.2
docling-parse==3.4.0
docx2txt==0.8
easyocr==1.7.2
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.17.0
filetype==1.2.0
Flask==3.0.1
flatbuffers==25.2.10
fonttools==4.56.0
frozendict @ file:///C:/b/abs_2alamqss6p/croot/frozendict_1713194885124/work
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
GPUtil==1.4.0
gradio==4.44.1
gradio_client==1.3.0
greenlet==3.1.1
grpcio==1.71.0rc2
grpcio-status==1.63.0rc1
gTTS==2.5.4
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
humanfriendly==10.0
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
imageio==2.33.1
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.1.2
Jinja2==3.1.3
jiter==0.8.2
joblib==1.4.2
jsonlines==3.1.0
jsonpatch @ file:///C:/b/abs_4fdm88t7zi/croot/jsonpatch_1714483974578/work
jsonpointer==2.1
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.3.1
lancedb==0.13.0
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
langcodes==3.5.0
langsmith==0.3.8
language_data==1.3.0
latex2mathml==3.77.0
lazy_loader==0.4
libmambapy @ file:///C:/b/abs_627vsv8bhu/croot/mamba-split_1734469608328/work/libmambapy
litellm==1.61.8
llvmlite==0.43.0
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py @ file:///C:/b/abs_a5bfngz6fu/croot/markdown-it-py_1684279915556/work
marko==2.1.2
MarkupSafe==2.1.4
marshmallow==3.26.1
matplotlib==3.4.2
mdurl @ file:///C:/Windows/TEMP/abs_3197pzpjbi/croots/recipe/mdurl_1659716032440/work
mediapipe==0.10.9
menuinst @ file:///C:/b/abs_fblttj5gp1/croot/menuinst_1738943438301/work
mpire==2.10.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
mypy-extensions==1.0.0
narwhals==1.28.0
nest-asyncio==1.6.0
networkx==2.5.1
ninja==1.11.1.3
nltk==3.9.1
numba==0.60.0
numpy==1.24.2
olefile==0.46
onnxruntime==1.19.2
openai==1.64.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.9.0.80
openpyxl==3.1.5
orjson==3.10.15
overrides==7.7.0
packaging==23.2
pafy==0.5.5
pandas==2.2.3
pathlib_abc==0.1.1
pathy==0.11.0
pdf2image==1.17.0
pdfminer.six==20231228
pdfminer3==2018.12.3.0
pdfplumber==0.11.4
peft==0.14.0
pillow==11.3.0
plac==1.1.3
platformdirs @ file:///C:/b/abs_b6z_yqw_ii/croot/platformdirs_1692205479426/work
plotly==6.0.0
pluggy @ file:///C:/b/abs_dfec_m79vo/croot/pluggy_1733170145382/work
pooch==1.8.2
preshed==3.0.9
propcache==0.2.1
proto-plus==1.26.0
protobuf==3.20.3
psutil==7.0.0
py==1.11.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycodestyle==2.7.0
pycosat @ file:///C:/b/abs_18nblzzn70/croot/pycosat_1736868434419/work
pycparser==2.20
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments @ file:///C:/b/abs_fay9dpq4n_/croot/pygments_1684279990574/work
pylance==0.17.0
PyMatting==1.1.13
PyMuPDF==1.25.3
PyMySQL==1.1.1
pyparsing==2.4.7
pypdf==5.3.0
pypdfium2==4.30.1
pypiwin32==223
pyreadline3==3.5.4
pyresparser==1.0.6
pyrsistent==0.20.0
PySocks @ file:///C:/ci/pysocks_1605307512533/work
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==1.0.2
pyttsx3==2.98
pytz==2021.1
PyWavelets==1.1.1
pywin32==308
PyYAML==6.0.2
ragas==0.2.13
referencing==0.36.2
regex==2024.11.6
rembg==2.0.54
requests==2.31.0
requests-toolbelt==1.0.0
retry==0.9.2
rich @ file:///C:/b/abs_8b92fzmygg/croot/rich_1732638991774/work
rpds-py==0.22.3
rsa==4.9
Rtree==1.3.0
ruamel.yaml @ file:///C:/b/abs_0cunwx_ww6/croot/ruamel.yaml_1727980181547/work
ruamel.yaml.clib @ file:///C:/b/abs_5fk8zi6n09/croot/ruamel.yaml.clib_1727769837359/work
ruff==0.9.6
safetensors==0.5.2
scikit-image==0.18.1
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
semchunk==2.2.2
sentence-transformers==3.4.1
shapely==2.0.7
shellingham==1.5.4
six==1.15.0
smart-open==6.4.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.6
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
sqlparse==0.4.1
srsly==2.5.1
starlette==0.45.3
streamlit==1.32.0
streamlit-tags==1.2.8
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
thinc==8.1.0
threadpoolctl==3.5.0
tifffile==2021.4.8
tiktoken==0.7.0
timm==1.0.14
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.12.0
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu
tornado==6.4.2
tqdm @ file:///C:/b/abs_0eh9b6xugj/croot/tqdm_1738945553987/work
transformers==4.47.1
typer==0.7.0
typing-inspect==0.9.0
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata==2025.1
ultralytics==8.3.80
ultralytics-thop==2.0.14
uritemplate==4.1.1
urllib3 @ file:///C:/b/abs_7bst06lizn/croot/urllib3_1737133657081/work
uvicorn==0.34.0
# Editable install with no version control (varag==0.1.0)
-e y:\final yr pro\varag
wasabi==0.10.1
watchdog==6.0.0
weasel==0.4.1
websockets==12.0
Werkzeug==3.0.1
whitenoise==6.11.0
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wrapt==1.17.2
XlsxWriter==3.2.2
xxhash==3.5.0
yarl==1.18.3
youtube-dl==2021.12.17
yt-dlp==2025.2.19
zipp==3.21.0
zstandard @ file:///C:/b/abs_31t8xmrv_h/croot/zstandard_1731356578015/work
